feat: allow for second, minute, and hour frequency
Start for freq finalization. In the long one should most likely separate
between the task index interval and the frequency used for the modelling
that use the frequency of the ts object. Ideally, the freq should be
more general and behave like in pandas/polars, i.e. accept an integer
before the interval string. Here we could support everything support by
seq.Date and seq.POSIXct and create an assertion for the possible
options.